move Group resource to dbt/artifacts and rename BaseArtifactNode to BaseResource
