+++ .changes/unreleased/Under the Hood-20240125-095453.yaml
@@ -1,6 +1,6 @@
 kind: Under the Hood
-body: Define Macro resource in dbt/artifacts
+body: Define Macro and Group resources in dbt/artifacts
 time: 2024-01-25T09:54:53.974332-05:00
 custom:
   Author: michelleark
-  Issue: "9382"
+  Issue: 9381 9382
