+++ core/dbt/artifacts/resources/v1/group.py
@@ -3,9 +3,11 @@ from typing import Literal
 
 from dbt.artifacts.resources.base import BaseResource
 from dbt.artifacts.resources.types import NodeType
+from dbt.artifacts.resources.v1.owner import Owner
 
 
 @dataclass
-class Documentation(BaseResource):
-    resource_type: Literal[NodeType.Documentation]
-    block_contents: str
+class Group(BaseResource):
+    name: str
+    owner: Owner
+    resource_type: Literal[NodeType.Group]
